Organize models and mock data
