+++ app/src/main/java/com/it235/nureserved/data/model/UserModel.kt
@@ -1,0 +1,10 @@
+package com.it235.nureserved.data.model
+
+data class User(
+    val firstName: String = "",
+    val middleName: String = "",
+    val lastName: String = "",
+    val email: String = "",
+    val program: String = "",
+    val studentNumber: String = ""
+)
